+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Create `NavTabCarousel` component.
 - Fix: Fix some styles on `SimpleCard`.
 - Fix: Remove `SearchInput` results cross.